Move local components to _components folder
