+++ src/app/_components/LatestIssues.tsx
@@ -1,7 +1,7 @@
+import { IssueStatusBadge } from '@/components'
 import prisma from '@/prisma/client'
 import { Avatar, Card, Flex, Heading, Table } from '@radix-ui/themes'
 import Link from 'next/link'
-import { IssueStatusBadge } from '../components'
 
 const LatestIssues = async () => {
   const latestIssues = await prisma.issue.findMany({
