+++ src/app/issues/[id]/_components/IssueDetails.tsx
