+++ src/app/issues/[id]/page.tsx
@@ -4,10 +4,10 @@ import { Box, Flex, Grid } from '@radix-ui/themes';
 import { getServerSession } from 'next-auth';
 import { notFound } from 'next/navigation';
 import { cache } from 'react';
-import AssigneeSelect from './AssigneeSelect';
-import DeleteIssueButton from './DeleteIssueButton';
-import EditIssueButton from './EditIssueButton';
-import IssueDetails from './IssueDetails';
+import AssigneeSelect from './_components/AssigneeSelect';
+import DeleteIssueButton from './_components/DeleteIssueButton';
+import EditIssueButton from './_components/EditIssueButton';
+import IssueDetails from './_components/IssueDetails';
 
 type Props = {
   params: {
