+++ src/app/issues/list/loading.tsx
@@ -1,6 +1,6 @@
 import { Skeleton } from '@/components';
 import { Table } from '@radix-ui/themes';
-import IssueActions from './IssueActions';
+import IssueActions from './_components/IssueActions';
 
 const LoadingIssuesPage = () => {
   const issues = [1, 2, 3, 4, 5];
