+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import { Container, Theme } from '@radix-ui/themes';
 import '@radix-ui/themes/styles.css';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import NavBar from './NavBar';
+import NavBar from './_components/NavBar';
 import './globals.css';
 import './theme-config.css';
