+++ src/app/page.tsx
@@ -1,9 +1,9 @@
 import prisma from "@/prisma/client";
 import { Flex, Grid } from "@radix-ui/themes";
 import { Metadata } from "next";
-import IssueChart from "./IssueChart";
-import IssueSummary from "./IssueSummary";
-import LatestIssues from "./LatestIssues";
+import IssueChart from "./_components/IssueChart";
+import IssueSummary from "./_components/IssueSummary";
+import LatestIssues from "./_components/LatestIssues";
 
 export default async function Home() {
   const open = await prisma.issue.count({ where: { status: 'OPEN' } });
